+++ CMakeLists.txt
@@ -74,11 +74,12 @@ set(MODMESH_HEADERS
     include/modmesh/base.hpp
     include/modmesh/profile.hpp
     include/modmesh/grid.hpp
-    include/modmesh/mesh.hpp
+    include/modmesh/mesh/mesh.hpp
     include/modmesh/mesh/StaticMesh.hpp
-    include/modmesh/small_vector.hpp
-    include/modmesh/ConcreteBuffer.hpp
-    include/modmesh/SimpleArray.hpp
+    include/modmesh/buffer/buffer.hpp
+    include/modmesh/buffer/small_vector.hpp
+    include/modmesh/buffer/ConcreteBuffer.hpp
+    include/modmesh/buffer/SimpleArray.hpp
 )
 string(REPLACE "include/" "${CMAKE_CURRENT_SOURCE_DIR}/include/"
        MODMESH_HEADERS "${MODMESH_HEADERS}")
